setup-apv: Support fetching vendor for a single device

+++ scripts/setup-apv.sh
@@ -6,13 +6,33 @@ SCRIPTPATH="$(cd "$(dirname "$0")";pwd -P)"
 source $SCRIPTPATH/metadata
 
 APV=vendor/android-prepare-vendor
+ALL_DEVICES=${devices[@]}
 
-for device in ${devices[@]}; do
+error() {
+  echo error: $1, please try again >&2
+  echo "Usage: $0 device"
+  echo "       Supported devices: ${ALL_DEVICES}"
+  echo "       Pass 'all' for all supported devices"
+  exit 1
+}
+
+[[ $# -ne 1 ]] && error "incorrect number of arguments"
+
+APV_DEVICE=$1
+
+if [[ ${APV_DEVICE} == "all" ]]; then
+	apv_devices=${ALL_DEVICES}
+else
+	apv_devices=${APV_DEVICE}
+fi
+
+for device in ${apv_devices}; do
 	echo "Setting up vendor for $device ${buildid[$device]}"
+	echo "Running $APV/execute-all.sh -y --debugfs -d $device -b ${buildid[$device]} -o $APV"
 	$APV/execute-all.sh -y --debugfs -d $device -b ${buildid[$device]} -o $APV || exit 1
 done
 
-for device in ${devices[@]}; do
+for device in ${apv_devices}; do
 	rm -rf vendor/google_devices/$device/*
 	cp -a $APV/$device/${buildid[$device]}/vendor/google_devices vendor/
 	rm -rf $APV/$device/${buildid[$device]}/vendor*
